pyproject.toml
[project]
name = "lio"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "query-loop", email = "dev.shubham.net@gmail.com"},
]
dependencies = ["fastapi>=0.115.4", "uvicorn>=0.32.0", "SQLAlchemy>=2.0.36", "pydantic>=2.9.2", "langchain>=0.3.7", "chromadb>=0.5.18", "dramatiq>=1.17.1", "redis>=5.2.0", "sqlmodel>=0.0.22", "sentence-transformers>=3.2.1", "qdrant-client>=1.12.1", "langchain-community>=0.3.4", "python-multipart>=0.0.17", "textract>=1.6.5", "PyPDF2>=3.0.1", "chardet>=3.0.4", "python-docx>=1.1.2", "openai>=1.54.3", "langchain-openai>=0.2.6", "python-magic>=0.4.27", "python-dotenv>=1.0.1", "cohere>=5.11.3"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
